+++ Runtime/ElementSelection/VisualModules/GameObjectToggleSelectionVisualModule.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 namespace UI.ElementSelection.VisualModules
 {
     [AddComponentMenu("UI/Element Selection/Modules/GameObject toggle SVM")]
-    public class GameObjectToggleSelectionVisualModule : MonoBehaviour, ISelectableElementVisualModule
+    public class GameObjectToggleSelectionVisualModule : StandaloneAnimatedSVM
     {
         [SerializeField] private GameObject _gameObject;
         [SerializeField] private SelectionModuleState<bool> _normalValue = new SelectionModuleState<bool> { Enabled = true, Value = false };
@@ -15,22 +15,47 @@ namespace UI.ElementSelection.VisualModules
         private bool _isHovered;
         private bool _isPremium;
         
-        public void OnSelectionChanged(bool isSelected)
+        private SelectionModuleState<bool>? _overriddenDefault;
+        private SelectionModuleState<bool>? _overriddenSelected;
+        private SelectionModuleState<bool>? _overriddenHovered;
+        private SelectionModuleState<bool>? _overriddenPremium;
+
+        public override void OnSelectionChanged(bool isSelected)
         {
             _isSelected = isSelected;
             UpdateActive();
         }
 
-        public void OnHoverChanged(bool isHovered)
+        public override void OnHoverChanged(bool isHovered)
         {
             _isHovered = isHovered;
             UpdateActive();
         }
 
-        public void OnPremium(bool isInPremiumState)
+        public override void OnPremium(bool isInPremiumState)
         {
             _isPremium = isInPremiumState;
             UpdateActive();
+        }
+
+        public override void SetToggleOverride(SelectionVisualState state, bool enabled, bool value)
+        {
+            SelectionModuleState<bool> overrideState = new SelectionModuleState<bool> { Enabled = enabled, Value = value };
+            switch (state)
+            {
+                case SelectionVisualState.Default: _overriddenDefault = overrideState; break;
+                case SelectionVisualState.Selected: _overriddenSelected = overrideState; break;
+                case SelectionVisualState.Hovered: _overriddenHovered = overrideState; break;
+                case SelectionVisualState.Premium: _overriddenPremium = overrideState; break;
+            }
+        }
+
+        public override void ClearOverrides()
+        {
+            _overriddenDefault = null;
+            _overriddenSelected = null;
+            _overriddenHovered = null;
+            _overriddenPremium = null;
         }
 
         private void UpdateActive()
@@ -39,17 +64,17 @@ namespace UI.ElementSelection.VisualModules
 
             bool targetActive;
 
-            if (_isPremium && _premiumValue.Enabled)
+            if (_isPremium && (_overriddenPremium ?? _premiumValue).Enabled)
             {
-                targetActive = _premiumValue.Value;
+                targetActive = (_overriddenPremium ?? _premiumValue).Value;
             }
             else
             {
-                targetActive = _isSelected ? _selectedValue.GetValue(true) : _normalValue.GetValue(false);
+                targetActive = _isSelected ? (_overriddenSelected?.GetValue(_normalValue.Value) ?? _selectedValue.GetValue(true)) : (_overriddenDefault?.GetValue(_normalValue.Value) ?? _normalValue.GetValue(false));
                 
-                if (_isHovered && _hoveredValue.Enabled)
+                if (_isHovered && (_overriddenHovered ?? _hoveredValue).Enabled)
                 {
-                    targetActive = _hoveredValue.Value;
+                    targetActive = (_overriddenHovered ?? _hoveredValue).Value;
                 }
             }